refactor: move Target,Scroe classes To Domain Package
