+++ app/src/main/java/org/slipp/masil/games/domains/highrow/Target.java
@@ -1,4 +1,4 @@
-package org.slipp.masil.games.domains;
+package org.slipp.masil.games.domains.highrow;
 
 import lombok.Value;
 import org.springframework.data.relational.core.mapping.Column;
