+++ app/src/test/java/org/slipp/masil/games/domains/highrow/HighLowJudgeTest.java
@@ -2,7 +2,6 @@ package org.slipp.masil.games.domains.highrow;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slipp.masil.games.domains.Target;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.slipp.masil.games.domains.highrow.HighLowJudgement.*;
