+++ app/src/test/java/org/slipp/masil/games/domains/highrow/HighLowPlayingContextRepositoryTest.java
@@ -1,15 +1,12 @@
 package org.slipp.masil.games.domains.highrow;
 
 import org.junit.jupiter.api.Test;
-import org.slipp.masil.games.domains.Target;
 import org.slipp.masil.games.domains.game.GameId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.annotation.DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD;
