+++ app/src/test/java/org/slipp/masil/games/domains/highrow/HighLowPlayingContextTest.java
@@ -2,8 +2,6 @@ package org.slipp.masil.games.domains.highrow;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slipp.masil.games.domains.Score;
-import org.slipp.masil.games.domains.Target;
 import org.slipp.masil.games.domains.game.GameId;
 
 import java.time.LocalDateTime;
