+++ app/src/test/java/org/slipp/masil/games/domains/ranking/RankingRepositoryTest.java
@@ -2,7 +2,7 @@ package org.slipp.masil.games.domains.ranking;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slipp.masil.games.domains.Score;
+import org.slipp.masil.games.domains.highrow.Score;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
